Fixed document details and editing bugs
Fixed quotes details and editing as well as the twig partials related to detail and editing.
****Removed deposit as a built-in custom field type and treated it as any other value.

+++ src/data_transfer_objects/TransferObject.php
@@ -45,7 +45,7 @@ class TransferObject
                 
                 $value = match ($type) {
                     'int' => (int)$value,
-                    'float' => (int)$value,
+                    'float' => (float)$value,
                     'string' => (string)$value,
                     'bool' => (bool)$value,
                     default => $value
